간이레시피 null 처리 수정

+++ app/src/main/java/com/example/dldke/foodbox/Adapter/HalfRecipeRecipeAdapter.java
@@ -31,8 +31,8 @@ public class HalfRecipeRecipeAdapter extends RecyclerView.Adapter<HalfRecipeReci
     @Override
     public void onBindViewHolder(ItemViewHolder holder, int position) {
         holder.mName.setText(mItems.get(position).getName());
-        int iCount = mItems.get(position).getCount();
-        String strCount = Integer.toString(iCount);
+        Double iCount = mItems.get(position).getCount();
+        String strCount = Double.toString(iCount);
 
         holder.mCount.setText(strCount);
     }
